Moved ToggleSwitch to components
