+++ src/main/java/com/ccs/baby/components/ToggleSwitch.java
@@ -1,0 +1,64 @@
+package com.ccs.baby.components;
+
+import javax.swing.*;
+import java.awt.*;
+import java.io.*;
+
+/**
+ * A custom JCheckBox that simulates a toggle switch with custom icons.
+ */
+public class ToggleSwitch extends JCheckBox {
+
+    public ToggleSwitch(String textValue, int horizontalPos, int verticalPos) {
+
+        ImageIcon downIcon = loadImage("/images/toggle_switch_down.png");
+        ImageIcon upIcon = loadImage("/images/toggle_switch_up.png");
+        ImageIcon holeIcon = loadImage("/images/hole.gif");
+
+        setIcon(upIcon);
+        setSelectedIcon(downIcon);
+        setDisabledIcon(holeIcon);
+
+        setFocusPainted(false);
+        setBorderPainted(false);
+        setContentAreaFilled(false);
+
+        setSelected(false);
+
+        setHorizontalTextPosition(horizontalPos);
+        setVerticalTextPosition(verticalPos);
+
+        setText(textValue);
+    }
+
+    protected ImageIcon loadImage(String image) {
+        int MAX_IMAGE_SIZE = 12400;  //Change this to the size of
+        //your biggest image, in bytes.
+        int count = 0;
+        BufferedInputStream imgStream = new BufferedInputStream(
+                this.getClass().getResourceAsStream(image));
+        if (imgStream != null) {
+            byte buf[] = new byte[MAX_IMAGE_SIZE];
+            try {
+                count = imgStream.read(buf);
+                imgStream.close();
+            } catch (java.io.IOException ioe) {
+                System.err.println("Couldn't read stream from file: " + image);
+                return null;
+            }
+            if (count <= 0) {
+                System.err.println("Empty file: " + image);
+                return null;
+            }
+            return new ImageIcon(Toolkit.getDefaultToolkit().createImage(buf));
+        } else {
+            System.err.println("Couldn't find file: " + image);
+            return null;
+        }
+    }
+
+}
+
+
+
+
